Refactor: use generic type in openModal, instead of useModal parameter
